# Server/api/openapi.yaml
...
openapi: "3.1.0"
info:
  title: "DictaDuo API"
  version: "2.0.0"
  description: "Shared transport contract for the Swift client and TypeScript server. Optional properties are omitted rather than encoded as null. Dates use whole-second ISO8601 UTC timestamps."
servers:
  -
    url: "http://localhost:8391"
security:
  -
    bearerAuth: []
...
paths:
...
  "/v2/captures":
    post:
      operationId: "startCapture"
      parameters:
        - "$ref": "#/components/parameters/CaptureOwner"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/StartCaptureRequest"
      responses:
        "201":
          description: "Admitted recording session with acknowledged microphone readiness."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/RecordingSnapshot"
        default:
          "$ref": "#/components/responses/APIError"
  "/v2/recordings/{id}/capture/heartbeat":
    post:
      operationId: "heartbeatCapture"
      parameters:
        - "$ref": "#/components/parameters/CaptureOwner"
        - name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
      responses:
        "204":
          description: "Owner lease renewed. Send every second; expiry is six seconds."
        default:
          "$ref": "#/components/responses/APIError"
  "/v2/recordings/{id}/capture/stop":
    post:
      operationId: "stopCapture"
      parameters:
        - "$ref": "#/components/parameters/CaptureOwner"
        - name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/StopCaptureRequest"
      responses:
        "202":
          description: "Capture stopped and its run sealed; processing continues to completion."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/RecordingSnapshot"
        default:
          "$ref": "#/components/responses/APIError"
  "/v2/recordings/{id}/context":
    post:
      operationId: "setCaptureContext"
      description: "Fix a remote capture's continuation (or none) before its text processing starts."
      parameters:
        - "$ref": "#/components/parameters/CaptureOwner"
        - name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/StopCaptureRequest"
      responses:
        "200":
          description: "Continuation fixed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/RecordingSnapshot"
        default:
          "$ref": "#/components/responses/APIError"
  "/v2/recordings/{id}/events":
    get:
      operationId: "recordingEvents"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
      responses:
        "200":
          description: "NDJSON RecordingSnapshot lines until the session completes, fails, or is discarded."
          content:
            "application/x-ndjson":
              schema:
                type: string
        default:
          "$ref": "#/components/responses/APIError"
...
  "/v1/generations/{id}/audio/{kind}":
    post:
      operationId: "appendAudio"
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/AudioChunkReceipt"
        default:
          "$ref": "#/components/responses/APIError"
      parameters:
        -
          name: "id"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/UUID"
        -
          name: "kind"
          in: "path"
          required: true
          schema:
            "$ref": "#/components/schemas/AudioKind"
        -
          name: "sequence"
          in: "query"
          required: true
          schema:
            type: "integer"
            minimum: 0
            maximum: 4095
        -
          name: "sampleRate"
          in: "query"
          required: true
          schema:
            type: "integer"
            minimum: 8000
            maximum: 192000
        -
          name: "channels"
          in: "query"
          required: true
          schema:
            type: "integer"
            minimum: 1
            maximum: 8
      requestBody:
        required: true
        content:
          "application/octet-stream":
            schema:
              type: "string"
              format: "binary"
        description: "Raw interleaved little-endian float32 PCM. Chunks must contain whole frames. Repeated sequences must have identical bytes."
        x-maximumBytes: 1048576
...
  /v2/recordings/capabilities:
    get:
      operationId: getRecordingCapabilities
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingCapabilities"
        default:
          $ref: "#/components/responses/APIError"
  /v2/recordings:
    post:
      operationId: createRecording
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingSnapshot"
        default:
          $ref: "#/components/responses/APIError"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGenerationRequest"
    get:
      operationId: listRecordings
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingPage"
        default:
          $ref: "#/components/responses/APIError"
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: before
          in: query
          schema:
            type: string
  /v2/recordings/{id}:
    get:
      operationId: getRecording
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingDetail"
        default:
          $ref: "#/components/responses/APIError"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
  /v2/recordings/{id}/transcript:
    get:
      operationId: getRecordingTranscript
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                type: string
        default:
          $ref: "#/components/responses/APIError"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
  /v2/recordings/{id}/audio/{kind}:
    get:
      operationId: getRecordingAudio
      responses:
        "200":
          description: Success
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        default:
          $ref: "#/components/responses/APIError"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: kind
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/AudioKind"
  /v2/recordings/{id}/discard:
    post:
      operationId: discardRecording
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingSnapshot"
        default:
          $ref: "#/components/responses/APIError"
      parameters:
        - "$ref": "#/components/parameters/CaptureMutationOwner"
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
  /v2/recordings/{id}/retry:
    post:
      operationId: retryRecording
      description: Resume a failed recording's processing from its committed text with local recognition. Nothing is delivered.
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingSnapshot"
        default:
          $ref: "#/components/responses/APIError"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
...
  /v2/recordings/{id}/stream:
    get:
      operationId: streamRecording
      description: Upgrade to WebSocket subprotocol dictaduo.recording.v1. Binary audio is UInt32BE JSON
        header length, UTF8 RecordingAudioHeader, then little-endian float32 PCM. Text controls
        resume, context, pause, stop and ping; server sends snapshots, acknowledgments, progress and
        errors. See docs/recording-protocol.md for durable ACK, fencing and resume semantics.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "101":
          description: WebSocket upgrade
        default:
          $ref: "#/components/responses/APIError"
  /v2/recordings/{id}/audio/{kind}/{runID}:
    get:
      operationId: getRecordingRunAudio
      responses:
        "200":
          description: Success
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        default:
          $ref: "#/components/responses/APIError"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: kind
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/AudioKind"
        - name: runID
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
components:
  schemas:
    UUID:
      type: "string"
      format: "uuid"
    GenerationMode:
      type: "string"
      enum:
        - "dictation"
        - "test"
        - "file"
...
    AudioKind:
      type: "string"
      enum:
        - "inference"
        - "original"
...
    DeviceIdentity:
      type: "object"
      additionalProperties: false
      properties:
        id:
          type: "string"
        name:
          type: "string"
      required:
        - "id"
        - "name"
...
    PreferencesSnapshot:
      type: "object"
      additionalProperties: false
      properties:
        revision:
          type: "integer"
          minimum: 0
          maximum: 9007199254740991
        preferences:
          "$ref": "#/components/schemas/ServerPreferences"
      required:
        - "revision"
        - "preferences"
...
    AudioSourceIdentity:
      type: "object"
      additionalProperties: false
      properties:
        hostID:
          type: "string"
          minLength: 1
          maxLength: 128
        id:
          type: "string"
          minLength: 1
          maxLength: 256
      required: [hostID, id]
...
    StartCaptureRequest:
      type: "object"
      additionalProperties: false
      properties:
        buttonTicket:
          "$ref": "#/components/schemas/UUID"
        requestID:
          "$ref": "#/components/schemas/UUID"
        device:
          "$ref": "#/components/schemas/DeviceIdentity"
        mode:
          type: "string"
          enum: [dictation, test]
        source:
          "$ref": "#/components/schemas/AudioSourceIdentity"
      required: [requestID, device, mode, source]
    StopCaptureRequest:
      type: "object"
      additionalProperties: false
      properties:
        continuationID:
          "$ref": "#/components/schemas/UUID"
    CreateGenerationRequest:
      type: "object"
      additionalProperties: false
      properties:
        requestID:
          "$ref": "#/components/schemas/UUID"
        device:
          "$ref": "#/components/schemas/DeviceIdentity"
        mode:
          "$ref": "#/components/schemas/GenerationMode"
      required:
        - "requestID"
        - "device"
        - "mode"
...
    AudioChunkReceipt:
      type: "object"
      additionalProperties: false
      properties:
        nextSequence:
          type: "integer"
          minimum: 0
          maximum: 9007199254740991
        frameCount:
          type: "integer"
          minimum: 0
          maximum: 9007199254740991
          format: "int64"
      required:
        - "nextSequence"
        - "frameCount"
...
    RecordingCapabilities:
      type: object
      additionalProperties: false
      properties:
        protocol:
          type: string
          enum:
            - dictaduo.recording.v1
        maximumPCMBytes:
          type: integer
          minimum: 0
          maximum: 9007199254740991
      required:
        - protocol
        - maximumPCMBytes
    RecordingRunEndpoint:
      type: object
      additionalProperties: false
      properties:
        runID:
          $ref: "#/components/schemas/UUID"
        inferenceFrames:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        originalFrames:
          type: integer
          minimum: 0
          maximum: 9007199254740991
      required:
        - runID
        - inferenceFrames
    RecordingStreamCheckpoint:
      type: object
      additionalProperties: false
      properties:
        runID:
          $ref: "#/components/schemas/UUID"
        kind:
          $ref: "#/components/schemas/AudioKind"
        format:
          $ref: "#/components/schemas/AudioStreamFormat"
        nextSequence:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        frameCount:
          type: integer
          minimum: 0
          maximum: 9007199254740991
      required:
        - runID
        - kind
        - format
        - nextSequence
        - frameCount
    RecordingSnapshot:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        requestID:
          $ref: "#/components/schemas/UUID"
        device:
          $ref: "#/components/schemas/DeviceIdentity"
        mode:
          $ref: "#/components/schemas/GenerationMode"
        settings:
          $ref: "#/components/schemas/PreferencesSnapshot"
        createdAt:
          type: string
          format: date-time
        revision:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        captureState:
          type: string
          enum:
            - recording
            - interrupted
            - stopped
            - discarded
        processingState:
          type: string
          enum:
            - queued
            - processing
            - completed
            - failed
        uploadedFrames:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        transcribedFrames:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        proofreadFrames:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        streams:
          type: array
          items:
            $ref: "#/components/schemas/RecordingStreamCheckpoint"
        epoch:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        stopRuns:
          type: array
          items:
            $ref: "#/components/schemas/RecordingRunEndpoint"
        error:
          type: string
        capture:
          $ref: "#/components/schemas/RemoteCapture"
        recognition:
          $ref: "#/components/schemas/RecognitionState"
        previewText:
          type: string
        continuationID:
          $ref: "#/components/schemas/UUID"
        closedRuns:
          type: array
          items:
            $ref: "#/components/schemas/RecordingRunEndpoint"
        runTimings:
          type: array
          items:
            $ref: "#/components/schemas/RecordingRunTiming"
      required:
        - id
        - requestID
        - device
        - mode
        - settings
        - createdAt
        - revision
        - captureState
        - processingState
        - uploadedFrames
        - transcribedFrames
        - proofreadFrames
        - streams
        - epoch
        - previewText
    RecordingDetail:
      type: object
      additionalProperties: false
      properties:
        snapshot:
          $ref: "#/components/schemas/RecordingSnapshot"
        result:
          $ref: "#/components/schemas/GenerationRecord"
      required:
        - snapshot
    RecordingPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecordingSnapshot"
        nextCursor:
          type: string
      required:
        - items
...
  parameters:
...
    CaptureMutationOwner:
      name: "X-DictaDuo-Capture-Owner"
      in: "header"
      required: false
      description: "Required for remote-capture cancellation and delivery; omitted by local-upload clients."
      schema:
        type: "string"
        pattern: "^[0-9a-f]{64}$"
    CaptureOwner:
      name: "X-DictaDuo-Capture-Owner"
      in: "header"
      required: true
      description: "Client-generated 256-bit lowercase hexadecimal secret, unique per capture request. Required in addition to server authorization for remote recording control and delivery. Never put it in URLs or history."
      schema:
        type: "string"
        pattern: "^[0-9a-f]{64}$"
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      description: "Optional on loopback. Required for remote binding. Native clients send no Origin header."
  responses:
    APIError:
      description: "A request, admission, or server error."
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/APIErrorResponse"